apps/api/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
motor>=3.3.2
pymongo>=4.6.0
beanie>=1.24.0
redis>=5.0.1

# HTTP client
httpx[http2]>=0.25.2
aiohttp>=3.9.5
urllib3>=2.2.0  # Force secure version (2.2.x series includes security fixes)

# Authentication
python-jose[cryptography]>=3.3.0
passlib[argon2,bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0

# Logging and monitoring
structlog>=23.2.0

# OpenTelemetry
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0
opentelemetry-instrumentation-logging>=0.42b0
opentelemetry-instrumentation-pymongo>=0.42b0
opentelemetry-instrumentation-redis>=0.42b0
opentelemetry-exporter-jaeger-thrift>=1.21.0
opentelemetry-exporter-otlp>=1.21.0

# Streaming
sse-starlette>=1.8.2

# Object Storage (V2 Future - Optional)
minio>=7.2.0

# Document processing (V1 - PDF text extraction)
pypdf>=3.17.0
pymupdf>=1.24.0  # PDF rasterization for image-only PDFs (scanned documents)
saptiva-agents>=0.2.3  # Saptiva SDK for PDF extraction via Custom Tools (requires Python 3.10+)

# PDF Report Generation (COPILOTO-414)
reportlab>=4.0.0  # Professional PDF report generation
opencv-python-headless>=4.8.0  # Logo detection and template matching

# OCR (MVP-OCR: Image text extraction)
pytesseract>=0.3.10
Pillow>=10.0.0
pillow-heif>=0.13.0

# Text Analysis & Validation (COPILOTO-414)
rapidfuzz>=3.0.0  # Fuzzy string matching for compliance validation
PyYAML>=6.0.0  # YAML configuration parsing

# Metrics
prometheus-client>=0.19.0
